weigence/requirements-dev.txt
# ====================================
# DESARROLLO - Dependencias adicionales
# ====================================

# Instalar con: pip install -r requirements.txt -r requirements-dev.txt

# Recarga automática en desarrollo
livereload==2.7.1
tornado==6.5.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Linting y formateo
flake8==6.1.0
black==23.12.1
pylint==3.0.3

# Debugging
ipdb==0.13.13
